[DEV] customarduino-tone

+++ main/customarduino.h
@@ -3,6 +3,7 @@
 
 #include <arduino.h>
 
+
 volatile uint8_t *portToDDR(uint8_t port);
 volatile uint8_t *portToPORT(uint8_t port);
 volatile uint8_t *portToPIN(uint8_t port);
@@ -14,6 +15,6 @@ int digitalRead_custom(uint8_t pin);
 
 void analogWrite_custom(uint8_t pin, uint8_t mode); // 아날로그 출력이 가능한 핀만 가능함 
 int analogRead_custom(uint8_t pin);
-void tone_custom(uint8_t pin, uint16_t frequency, uint16_t duration);
+void tone_custom(uint8_t pin, unsigned int frequency, unsigned long duration);
 
 #endif
